Fix/ favorites courses method moved to Statics Service and favorites courses add on endpoint moved to Statics Controller
