+++ src/main/java/com/orange_evolution_backend/controller/CourseController.java
@@ -85,11 +85,7 @@ public class CourseController {
         return new ResponseEntity<Course>(courseService.saveCourse(course), HttpStatus.CREATED);
     }
 
-    @ApiOperation(value = "add a favorite course")
-    @GetMapping("/favorites/{userId}/{courseId}")
-    public ResponseEntity<Course> addFavoriteCourse(@PathVariable Long userId, @PathVariable Long courseId) {
-        return ResponseEntity.ok(courseService.favoriteCourse(userId, courseId));
-    }
+
 
     @ApiOperation(value = "Delete a course")
     @DeleteMapping("/{courseId}")
