+++ src/main/java/com/orange_evolution_backend/service/StatisticService.java
@@ -1,6 +1,7 @@
 package com.orange_evolution_backend.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
@@ -16,6 +17,23 @@ import lombok.AllArgsConstructor;
 public class StatisticService {
     CourseRepository courseRepository;
     UserRepository userRepository;
+
+
+    public Course favoriteCourse(Long idUser, Long idCourse){
+        Optional<User> userOpt = userRepository.findById(idUser);
+        Optional<Course> courseOpt = courseRepository.findById(idCourse);
+
+        if(userOpt.isPresent() && courseOpt.isPresent()){
+            Course course = courseOpt.get();
+            course.getUsers().add(userOpt.get());
+
+            return courseRepository.save(course);
+            
+        }
+
+        return null;
+    }
+    
     
     public List<Course> findFavoritesCoursesByIdUser(Long idUser, Long idcourse){
         User userFavorite = userRepository.findById(idcourse).get();
